Improve code refactoring

+++ src/tools/file.rs
@@ -1,7 +1,34 @@
-use std::{env, fmt::{Error, Result}, fs::{self, File}, io::{BufReader, Read, Seek, SeekFrom, Write}, path::Path};
+use std::{fs::{self, File, metadata}, io::{BufReader, Read, Seek, SeekFrom, Write}, path::Path};
 
-//const KB = 1024;
+use crate::common::errors;
 
+fn print_copy(filename: &str, total: &u64, current: &u64) {
+    // Move to beginning of previous line & Clear entire line
+    print!("\x1b[1F\x1b[2K");
+    let percent:f64 = (current * 100 / total) as f64;
+    println!("{}  {}  {}%", filename, get_size(total), percent);   
+}
+
+fn get_size(total: &u64)-> String {
+
+    let mut prefix = "bytes";
+    let mut unit = 1;
+    
+    if total > &crate::consts::GIGABYTE {
+        prefix = "gb";
+        unit = crate::consts::GIGABYTE;
+    }
+    else if total > &crate::consts::MEGABYTE {
+        prefix = "mb";
+        unit = crate::consts::MEGABYTE;
+    }
+    else if total > &crate::consts::KILOBYTE {
+        prefix = "kb";
+        unit = crate::consts::KILOBYTE;
+    }
+
+    String::from(format!("{:.3} {}", *total as f64 / unit as f64, prefix))
+}
 
 pub struct FileHelper{
     //source:String,
@@ -12,15 +39,25 @@ impl FileHelper{
     
     /**
      */
-    pub fn copy(source:&String, destination:&String, limit_bytes:&Option<u32>) -> Result {
+    pub fn copy(source:&str, destination:&str, limit_bytes:&Option<u64>) -> errors::AppResult<()> {
         // Take path
         let source_path = Path::new(source);
         let destination_path = Path::new(destination);
+
+        if source_path.is_dir() {
+            return Err(errors::AppError::new("Unsuported directory copy"));
+        }
+
+        if destination_path.is_dir() {
+            return Err(errors::AppError::new("Unsuported directory destination. please add a file name"));
+        }
         
-        let mut megas: u64 = 50 * crate::MEGABYTE as u64;
+        let filename = destination_path.file_name().unwrap().to_str().unwrap();
+
+        let mut megas: u64 = 50 * crate::consts::MEGABYTE as u64;
 
         if limit_bytes.is_some() {
-            megas = limit_bytes.unwrap() as u64;
+            megas = limit_bytes.unwrap();
         }
 
         // TODO: need to clear/create empty file, because then we're appending bytes to file
@@ -34,37 +71,34 @@ impl FileHelper{
         */
         FileHelper::clear_file(destination_path);
 
-
-        println!("The current directory is {:?}", source_path.join(&source) );
-
         let f = File::open(source_path).expect("no file found");
         let metadata = f.metadata().expect("File error");
         //dbg!(&metadata);
 
-        
-
         let mut total = metadata.len();
-        let mut cursorPos = 0;
-        let mut bytes =  0;
+        let mut cursor_pos = 0;
+        //let mut bytes =  0;
 
         while  total > 0{
-            bytes = megas;
+            let mut bytes = megas;
             if megas > total{
                 bytes = total;
             }
 
-            println!("Total bytes to copy:{}, Total copied bytes:{}, to copy:{}, ", total, &cursorPos, &bytes);
+            //println!("Total bytes to copy:{}, Total copied bytes:{}, to copy:{}, ", total, &cursor_pos, &bytes);
+            
 
             // Create buffer an move the starting position of the stream
             let mut buffer = vec![0; bytes as usize];
             let mut reader = BufReader::new(&f);
             // TODO: Add validation here to retry
-            reader.seek(SeekFrom::Start(cursorPos));
-            reader.read(&mut buffer);
+            reader.seek(SeekFrom::Start(cursor_pos)).unwrap();
+            reader.read(&mut buffer).unwrap();
 
-            cursorPos += bytes;
+            cursor_pos += bytes;
             total -= bytes;
-            println!("current copied: {}", cursorPos / (crate::MEGABYTE as u64));
+            
+            print_copy(filename, &metadata.len(), &cursor_pos);
             
             // TODO: this options are to append bytes to existing file. 
             let mut file = fs::OpenOptions::new()
@@ -82,9 +116,9 @@ impl FileHelper{
     /**
      * Delete file if exist or create a new empty file
      */
-    fn clear_file(path:&Path) {
+    pub fn clear_file(path:&Path) {
 
-        if path.exists() {
+        if path.exists() && path.is_file() {
             // Delete file
             fs::remove_file(path).unwrap();
         }
